These are just tiny adjustments that won't make the computations 10 times faster but that will at least optimize a bit the overall performance.

I run the following benchmark before and after each commit:

```rb
require "benchmark"

ephem = Astronoby::Ephem.load("tmp/de440s.bsp")

Benchmark.bmbm do |x|
  x.report { 
    50.times do
      utc_offset = "+01:00"
      date = Date.today
      observer = Astronoby::Observer.new(
        latitude: Astronoby::Angle.from_degrees(48.836606722281196),
        longitude: Astronoby::Angle.from_degrees(2.3363422188788454),
        elevation: Astronoby::Distance.from_meters(35),
        utc_offset: utc_offset
      )
      calculator = Astronoby::RisingTransitSettingEventsCalculator.new(
        observer: observer,
        target_body: Astronoby::Moon,
        ephem: ephem
      )
      events = calculator.events_on(date)
      events.rising_azimuth.str(:dms)
    end
  }
end
```

### Before PR

```
Rehearsal ------------------------------------
  74.885694   0.378245  75.263939 ( 75.450283)
-------------------------- total: 75.263939sec

       user     system      total        real
  74.567673   0.360459  74.928132 ( 74.991962)
```

### After freezing nutation terms

```
Rehearsal ------------------------------------
  73.914803   0.391628  74.306431 ( 74.442645)
-------------------------- total: 74.306431sec

       user     system      total        real
  73.768577   0.354093  74.122670 ( 74.189397)
```

### After implementing a better crossing finder

```
Rehearsal ------------------------------------
  67.089399   0.360070  67.449469 ( 67.622139)
-------------------------- total: 67.449469sec

       user     system      total        real
  66.947562   0.488231  67.435793 ( 67.482280)
```

### After pre-computing leap seconds

```
Rehearsal ------------------------------------
  66.408339   0.223447  66.631786 ( 66.716871)
-------------------------- total: 66.631786sec

       user     system      total        real
  66.519351   0.192138  66.711489 ( 66.753833)
```

### Conclusion

As you can see, this is not a game changer and the library can still be considered as pretty slow, there are other significant optimizations that will need to be implemented in the future, but each change made some little progress.